backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
alembic==1.16.4
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
apscheduler==3.11.0
pytesseract==0.3.13
Pillow==11.3.0
openai==1.97.1
pydantic==2.11.7
pydantic[email]==2.11.7
python-dateutil==2.9.0